+++ src/Indice.Features.Cases.UI/CHANGELOG.md
@@ -6,8 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
+- Categories & Ordering for case types
 - WYSIWYG widget, using `ngx-quill`
 
 ## [6.11.3] - 2023-03-15